ai_service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
mediapipe==0.10.5
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
